scmp: add payload size support to scmp echo
Also simplifies argument processing by refactoring scmp echo
to use cobra.

The remote address as a positional argument instead of a -remote flag.
For example, `scmp echo 64-2:0:a,[10.0.2.162]`.

`scmp` now uses POSIX style flags (prefixed with double dash, --)
instead of Go style flags (in other words, now use `--refresh` instead of
`-refresh`). scmp command lines that use Go style flags now have changed
semantics (e.g., `-refresh` actually means setting the `-r`, `-e`, `-f`,
etc. flags). Please ensure all scripts containing the old CLI flags have
been inspected and updated.

+++ go/tools/scmp/echo/echo.go
@@ -75,7 +75,7 @@ func sendPkts() {
 		}
 		cmn.Stats.Sent += 1
 		// More packets?
-		if cmn.Count != 0 && cmn.Stats.Sent == cmn.Count {
+		if cmn.Count != 0 && cmn.Stats.Sent == uint(cmn.Count) {
 			break
 		}
 		// Update packet fields
